refactor: move tracebackBag to new gnrerror module, drop unused errorTxt/errorLog
- Create gnr.core.gnrerror with tracebackBag and _is_library_frame
- Remove errorTxt and errorLog from gnrlog (confirmed unused)
- Remove orphan imports from gnrlog (GnrStructData, toText, sendmail, Bag, traceback, datetime, hashlib, linecache, json)
- Update all call sites: gnrapp.py, lgcy/transaction.py, gnrlang_test.py

Ref #710
